+++ lualib/http/tlshelper.lua
@@ -43,13 +43,16 @@ function tlshelper.closefunc(tls_ctx)
 end
 
 function tlshelper.readfunc(fd, tls_ctx)
-    local readfunc = socket.readfunc(fd)
+    local function readfunc()
+        readfunc = socket.readfunc(fd)
+        return ""
+    end
     local read_buff = ""
     return function (sz)
         if not sz then
             local s = ""
             if #read_buff == 0 then
-                local ds = readfunc(sz)
+                local ds = readfunc()
                 s = tls_ctx:read(ds)
             end
             s = read_buff .. s
@@ -77,7 +80,6 @@ function tlshelper.writefunc(fd, tls_ctx)
 end
 
 function tlshelper.readallfunc(fd, tls_ctx)
-    local readfunc = socket.readfunc(fd)
     return function ()
         local ds = socket.readall(fd)
         local s = tls_ctx:read(ds)
@@ -89,8 +91,8 @@ function tlshelper.newctx()
     return c.newctx()
 end
 
-function tlshelper.newtls(method, ssl_ctx)
-    return c.newtls(method, ssl_ctx)
+function tlshelper.newtls(method, ssl_ctx, hostname)
+    return c.newtls(method, ssl_ctx, hostname)
 end
 
 return tlshelper