+++ lualib/http/websocket.lua
@@ -45,6 +45,7 @@ local function write_handshake(self, host, url, header)
     if code ~= 101 then
         error(string.format("websocket handshake error: code[%s] info:%s", code, body))
     end
+	assert(body == "")	-- todo: M.read may need handle it
 
     if not recvheader["upgrade"] or recvheader["upgrade"]:lower() ~= "websocket" then
         error("websocket handshake upgrade must websocket")
@@ -183,7 +184,7 @@ local function write_frame(self, op, payload_data, masking_key)
     -- mask set to 0
     if payload_len < 126 then
         s = string.pack("I1I1", v1, mask | payload_len)
-    elseif payload_len < 0xffff then
+    elseif payload_len <= 0xffff then
         s = string.pack("I1I1>I2", v1, mask | 126, payload_len)
     else
         s = string.pack("I1I1>I8", v1, mask | 127, payload_len)
@@ -302,11 +303,10 @@ end
 
 
 local SSLCTX_CLIENT = nil
-local function _new_client_ws(socket_id, protocol)
+local function _new_client_ws(socket_id, protocol, hostname)
     local obj
     if protocol == "ws" then
         obj = {
-            websocket = true,
             close = function ()
                 socket.close(socket_id)
             end,
@@ -319,11 +319,10 @@ local function _new_client_ws(socket_id, protocol)
     elseif protocol == "wss" then
         local tls = require "http.tlshelper"
         SSLCTX_CLIENT = SSLCTX_CLIENT or tls.newctx()
-        local tls_ctx = tls.newtls("client", SSLCTX_CLIENT)
+        local tls_ctx = tls.newtls("client", SSLCTX_CLIENT, hostname)
         local init = tls.init_requestfunc(socket_id, tls_ctx)
         init()
         obj = {
-            websocket = true,
             close = function ()
                 socket.close(socket_id)
                 tls.closefunc(tls_ctx)()
@@ -435,17 +434,21 @@ function M.connect(url, header, timeout)
     end
 
     assert(host)
-    local host_name, host_port = string.match(host, "^([^:]+):?(%d*)$")
-    assert(host_name and host_port)
+    local host_addr, host_port = string.match(host, "^([^:]+):?(%d*)$")
+    assert(host_addr and host_port)
     if host_port == "" then
         host_port = protocol == "ws" and 80 or 443
     end
+    local hostname
+    if not host_addr:match(".*%d+$") then
+        hostname = host_addr
+    end
 
     uri = uri == "" and "/" or uri
-    local socket_id = sockethelper.connect(host_name, host_port, timeout)
-    local ws_obj = _new_client_ws(socket_id, protocol)
+    local socket_id = sockethelper.connect(host_addr, host_port, timeout)
+    local ws_obj = _new_client_ws(socket_id, protocol, hostname)
     ws_obj.addr = host
-    write_handshake(ws_obj, host_name, uri, header)
+    write_handshake(ws_obj, host_addr, uri, header)
     return socket_id
 end